requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
cachetools==5.2.0
    # via google-auth
certifi==2019.6.16
    # via requests
chardet==3.0.4
    # via requests
google-auth==2.12.0
    # via
    #   -r requirements.in
    #   google-auth-oauthlib
google-auth-oauthlib==0.5.3
    # via -r requirements.in
idna==2.8
    # via requests
oauthlib==3.2.1
    # via requests-oauthlib
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
requests==2.22.0
    # via
    #   -r requirements.in
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via google-auth-oauthlib
rsa==4.9
    # via google-auth
six==1.12.0
    # via google-auth
urllib3==1.25.3
    # via requests
